+++ HISTORY.md
@@ -4,6 +4,8 @@ History
 ### By now 2015-04-01
 * Add -v option for bin/biotcm
 * Add BioTCM::Interfaces
+* Add BioTCM::Layer
+* Deprecate Graph
 
 ### v0.2.2 2015-03-18
 * Fix a bug in BioTCM::Databases::OMIM